Override Rack::Rquest#tursted_proxy?
This is needed to allow `request.ip` to work in Sinatra in a way that
makes sense on internal networks. The default implementation will remove
RFC1918 addresses from being returned when in X-Forwarded-For. In a lot
of cases, simply removing 127.0.0.1 and localhost is more than enough.
If there's a demand, this could be made into a setting.