Change the way we handle record updates.
When a node goes live, the PDNS Listener will try to insert a record
into PDNS. If the IP already has a record, but there's no record for
hostname, currently it'll just do nothing.

This update will nuke the IP record,and create a new record in its place
with the correct information.